+++ previous-versions/files/RELEASES.md
@@ -309,3 +309,13 @@ Based on:
 - [go v0.24.0] previous-versions/files
 ### Releases
 - [Go v0.24.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/files/v0.24.0 - previous-versions/files
+
+## 2023-11-06 01:26:40
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.25.0] previous-versions/files
+### Releases
+- [Go v0.25.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/files/v0.25.0 - previous-versions/files